[7938] Use same type for account ids in different tables related to used in code uint32 type for it.
(cherry picked from commit 1167b8962732f2423852a6d2331fcf51af8dfa89)

This is seldom case table structure modificatiom that not required revert before upgrade to 0.13

+++ sql/realmd.sql
@@ -41,7 +41,7 @@ UNLOCK TABLES;
 
 DROP TABLE IF EXISTS `account`;
 CREATE TABLE `account` (
-  `id` bigint(20) unsigned NOT NULL auto_increment COMMENT 'Identifier',
+  `id` int(11) unsigned NOT NULL auto_increment COMMENT 'Identifier',
   `username`      varchar(32) NOT NULL default '',
   `sha_pass_hash` varchar(40) NOT NULL default '',
   `gmlevel` tinyint(3) unsigned NOT NULL default '0',
